app/requirements.txt
flask
openai
streamlit
requests
python-dotenv
azure-identity
azure-ai-openai